feat: add Generate Text "PERORANGAN"
